Challenge Topics

Challenge Topics

topic_tags

format_topics(params)

Arguments

params
Challenge parameters as list.

Format

topic_tags is a character vector of valid topic tags.

Value

format_topics() returns a character value indicating the challenge topic tags.

Examples

topic_tags

#>   [1] "aCGH"                           "adme16cod"                     
#>   [3] "AffymetrixChip"                 "ag"                            
#>   [5] "AgilentChip"                    "Agroinformatics"               
#>   [7] "Alignment"                      "AlternativeSplicing"           
#>   [9] "Annotation"                     "AnnotationData"                
#>  [11] "AnnotationHub"                  "AnnotationHubSoftware"         
#>  [13] "AnnotationWorkflow"             "Anopheles_gambiae"             
#>  [15] "Anopheles_gambiae_Data"         "Apis_mellifera"                
#>  [17] "Apis_mellifera_Data"            "Arabidopsis_lyrata"            
#>  [19] "Arabidopsis_lyrata_Data"        "Arabidopsis_thaliana"          
#>  [21] "Arabidopsis_thaliana_Data"      "ArrayExpress"                  
#>  [23] "Asparagus_officinalis"          "AssayDomain"                   
#>  [25] "AssayDomainData"                "ATACSeq"                       
#>  [27] "ath1121501"                     "Bacillus_subtilis"             
#>  [29] "Bacillus_subtilis_Data"         "BasicWorkflow"                 
#>  [31] "BatchEffect"                    "Bayesian"                      
#>  [33] "BioCarta"                       "BiocViews"                     
#>  [35] "BiologicalQuestion"             "BiomedicalInformatics"         
#>  [37] "Biophysics"                     "Bos_taurus"                    
#>  [39] "Bos_taurus_Data"                "BreastCancerData"              
#>  [41] "BSgenome"                       "Caenorhabditis_elegans"        
#>  [43] "Caenorhabditis_elegans_Data"    "Callithrix_jacchus"            
#>  [45] "Callithrix_jacchus_Data"        "CancerData"                    
#>  [47] "Canis_familiaris"               "Canis_familiaris_Data"         
#>  [49] "cdf"                            "celegans"                      
#>  [51] "CellBasedAssays"                "CellBiology"                   
#>  [53] "CellCulture"                    "CGHData"                       
#>  [55] "Challenges"                     "Cheminformatics"               
#>  [57] "ChIPchip"                       "ChipDb"                        
#>  [59] "ChipManufacturer"               "ChipName"                      
#>  [61] "ChipOnChip"                     "ChipOnChipData"                
#>  [63] "ChIPSeq"                        "ChIPSeqData"                   
#>  [65] "Chlamydomonas_reinhardtii"      "Cicer_arietinum"               
#>  [67] "Ciona_intestinalis"             "Ciona_intestinalis_Data"       
#>  [69] "Classification"                 "ClonetechChip"                 
#>  [71] "Clustering"                     "CodelinkChip"                  
#>  [73] "ColonCancerData"                "Community"                     
#>  [75] "ComparativeGenomics"            "ComputationalChemistry"        
#>  [77] "COPDData"                       "CopyNumberVariation"           
#>  [79] "CopyNumberVariationData"        "Coverage"                      
#>  [81] "CpGIsland"                      "CpGIslandData"                 
#>  [83] "CRISPR"                         "CustomArray"                   
#>  [85] "CustomCDF"                      "CustomDBSchema"                
#>  [87] "Danio_rerio"                    "Danio_rerio_Data"              
#>  [89] "DataImport"                     "DataRepresentation"            
#>  [91] "db0"                            "ddPCR"                         
#>  [93] "DecisionTree"                   "DemethylateRegionDetection"    
#>  [95] "DenovoAssembler"                "DenovoGenome"                  
#>  [97] "DenovoTranscriptome"            "DifferentialExpression"        
#>  [99] "DifferentialMethylation"        "DifferentialPeakCalling"       
#> [101] "DifferentialSplicing"           "DifferentialSplicingWorkflow"  
#> [103] "DimensionReduction"             "DiseaseModel"                  
#> [105] "DNA3DStructure"                 "DNAMethylation"                
#> [107] "DNASeq"                         "DNASeqData"                    
#> [109] "DNaseSeq"                       "DriverMutation"                
#> [111] "drosgenome1"                    "Drosophila_melanogaster"       
#> [113] "Drosophila_melanogaster_Data"   "Drosophila_virilis"            
#> [115] "Drosophila_virilis_Data"        "drosophila2"                   
#> [117] "ENCODE"                         "Epigenetics"                   
#> [119] "EpigeneticsWorkflow"            "Epitranscriptomics"            
#> [121] "Eremothecium_gossypii"          "Eremothecium_gossypii_Data"    
#> [123] "Escherichia_coli"               "Escherichia_coli_Data"         
#> [125] "EuPathDB"                       "ExomeSeq"                      
#> [127] "ExomeSeqData"                   "ExonArray"                     
#> [129] "ExperimentalDesign"             "ExperimentData"                
#> [131] "ExperimentHub"                  "ExperimentHubSoftware"         
#> [133] "ExpressionData"                 "FeatureExtraction"             
#> [135] "FlowCytometry"                  "FlowCytometryData"             
#> [137] "FRMA"                           "FunctionalAnnotation"          
#> [139] "FunctionalGenomics"             "FunctionalPrediction"          
#> [141] "GACustomCDF"                    "Gallus_gallus"                 
#> [143] "Gallus_gallus_Data"             "Gasterosteus_aculeatus"        
#> [145] "Gasterosteus_aculeatus_Data"    "GEChip"                        
#> [147] "GeneCardsCustomSchema"          "GeneExpression"                
#> [149] "GeneExpressionWorkflow"         "GeneFusionDetection"           
#> [151] "GenePrediction"                 "GeneRegulation"                
#> [153] "GeneSetEnrichment"              "GeneSignaling"                 
#> [155] "GeneTarget"                     "Genetics"                      
#> [157] "GeneticVariability"             "Genome"                        
#> [159] "GenomeAnnotation"               "GenomeAssembly"                
#> [161] "GenomeBrowsers"                 "GenomeWideAssociation"         
#> [163] "GenomicSequence"                "GenomicVariantsWorkflow"       
#> [165] "GenomicVariation"               "GenotypingArray"               
#> [167] "GenotypingArrayData"            "GEO"                           
#> [169] "Germline"                       "GermlineMutation"              
#> [171] "Glycine_max"                    "Glycine_max_Data"              
#> [173] "GO"                             "GraphAndNetwork"               
#> [175] "GUI"                            "h10kcod"                       
#> [177] "h20kcod"                        "HapMap"                        
#> [179] "hcg110"                         "hcgi12k"                       
#> [181] "hcgi8k"                         "hgfocus"                       
#> [183] "hgu133a"                        "hgu133a2"                      
#> [185] "hgu133b"                        "hgu133plus2"                   
#> [187] "hgu95a"                         "hgu95av2"                      
#> [189] "hgu95b"                         "hgu95c"                        
#> [191] "hgu95d"                         "hgu95e"                        
#> [193] "hguatlas13k"                    "hgug4100a"                     
#> [195] "hgug4101a"                      "hgug4110b"                     
#> [197] "hgug4111a"                      "hgug4112a"                     
#> [199] "hguqiagenv3"                    "hi16cod"                       
#> [201] "HiC"                            "HiddenMarkovModel"             
#> [203] "HighThroughputImagingData"      "HistoneModification"           
#> [205] "HIVData"                        "Homo_sapiens"                  
#> [207] "Homo_sapiens_Data"              "Hordeum_vulgare"               
#> [209] "Hordeum_vulgare_Data"           "hs25kresogen"                  
#> [211] "hu35ksuba"                      "hu35ksubb"                     
#> [213] "hu35ksubc"                      "hu35ksubd"                     
#> [215] "hu6800"                         "HuO22"                         
#> [217] "hwgcod"                         "IlluminaChip"                  
#> [219] "illuminaHumanv1"                "illuminaHumanv2"               
#> [221] "illuminaMousev1"                "illuminaMousev1p1"             
#> [223] "illuminaRatv1"                  "ImagingMassSpectrometry"       
#> [225] "ImagingMassSpectrometryData"    "ImmunoOncology"                
#> [227] "ImmunoOncologyData"             "ImmunoOncologyWorkflow"        
#> [229] "indac"                          "INDACChip"                     
#> [231] "IndelDetection"                 "Infrastructure"                
#> [233] "InparanoidDb"                   "Instrumentation"               
#> [235] "JazaerimetaData"                "KEGG"                          
#> [237] "KidneyCancerData"               "Kluyveromyces_lactis"          
#> [239] "Kluyveromyces_lactis_Data"      "LeukemiaCancerData"            
#> [241] "LinkageDisequilibrium"          "Lipidomics"                    
#> [243] "lumiHumanV1"                    "lumiHumanV2"                   
#> [245] "lumiMouseV1"                    "lumiRatV1"                     
#> [247] "LungCancerData"                 "m10kcod"                       
#> [249] "m20kcod"                        "Macaca_fascicularis"           
#> [251] "Macaca_mulatta"                 "Macaca_mulatta_Data"           
#> [253] "Magnaporthe_grisea"             "Magnaporthe_grisea_Data"       
#> [255] "MassSpectrometry"               "MassSpectrometryData"          
#> [257] "MathematicalBiology"            "MBNICustomCDF"                 
#> [259] "Medicago_truncatul_Data"        "Medicago_truncatula"           
#> [261] "MeSHDb"                         "Metabolomics"                  
#> [263] "MetagenomeAssembly"             "Metagenomics"                  
#> [265] "MethylationArray"               "MethylationArrayData"          
#> [267] "MethylSeq"                      "MethylSeqData"                 
#> [269] "mgu74a"                         "mgu74av2"                      
#> [271] "mgu74b"                         "mgu74bv2"                      
#> [273] "mgu74c"                         "mgu74cv2"                      
#> [275] "mguatlas5k"                     "mgug4121a"                     
#> [277] "mgug4122a"                      "mi16cod"                       
#> [279] "Microarray"                     "MicroarrayData"                
#> [281] "MicrobialStrain"                "Microbiome"                    
#> [283] "MicrobiomeData"                 "MicroRNAArray"                 
#> [285] "MicroRNAArrayData"              "MicrosatelliteDetection"       
#> [287] "MicrotitrePlateAssay"           "MicrotitrePlateAssayData"      
#> [289] "miRNA"                          "miRNAData"                     
#> [291] "mm24kresogen"                   "MNaseSeq"                      
#> [293] "moe430a"                        "moe430b"                       
#> [295] "Monodelphis_domestica"          "Monodelphis_domestica_Data"    
#> [297] "MotifAnnotation"                "MotifDiscovery"                
#> [299] "mouse4302"                      "mouse430a2"                    
#> [301] "mpedbarray"                     "mRNAArrayData"                 
#> [303] "mRNAMicroarray"                 "mu11ksuba"                     
#> [305] "mu11ksubb"                      "Mu15v1"                        
#> [307] "mu19ksuba"                      "mu19ksubb"                     
#> [309] "mu19ksubc"                      "Mu22v3"                        
#> [311] "MultiChannel"                   "MultiChannelData"              
#> [313] "MultidimensionalScaling"        "MultipleComparison"            
#> [315] "MultipleSequenceAlignment"      "Mus_musculus"                  
#> [317] "Mus_musculus_Data"              "mwgcod"                        
#> [319] "NCI"                            "NCINatureCurated"              
#> [321] "Network"                        "NetworkEnrichment"             
#> [323] "NetworkInference"               "NeuralNetwork"                 
#> [325] "Neurospora_crassa"              "Neurospora_crassa_Data"        
#> [327] "Normalization"                  "Norway981"                     
#> [329] "NucleosomePositioning"          "Oncorhynchus_mykiss"           
#> [331] "Oncorhynchus_mykiss_Data"       "OneChannel"                    
#> [333] "OneChannelData"                 "OperonHumanV3"                 
#> [335] "Organism"                       "OrganismData"                  
#> [337] "OrganismDb"                     "OrgDb"                         
#> [339] "Oryza_sativa"                   "Oryza_sativa_Data"             
#> [341] "OvarianCancerData"              "PackageType"                   
#> [343] "PackageTypeData"                "Pan_paniscus"                  
#> [345] "Pan_paniscus_Data"              "Pan_troglodytes"               
#> [347] "Pan_troglodytes_Data"           "PartheenMetaData"              
#> [349] "PathwayInteractionDatabase"     "Pathways"                      
#> [351] "PatternLogic"                   "PeakDetection"                 
#> [353] "pedbarrayv10"                   "pedbarrayv9"                   
#> [355] "Pharmacogenetics"               "Pharmacogenomics"              
#> [357] "Phylogenetics"                  "Plasmodium_falciparum"         
#> [359] "Plasmodium_falciparum_Data"     "PolyPhen"                      
#> [361] "PooledScreens"                  "Preprocessing"                 
#> [363] "PrincipalComponent"             "probe"                         
#> [365] "Project1000genomes"             "ProprietaryPlatforms"          
#> [367] "ProprietaryPlatformsData"       "ProstateCancerData"            
#> [369] "Proteome"                       "Proteomics"                    
#> [371] "ProteomicsWorkflow"             "Pseudomonas_aeruginosa"        
#> [373] "Pseudomonas_aeruginosa_Data"    "QiagenChip"                    
#> [375] "qPCR"                           "qPCRData"                      
#> [377] "QualityControl"                 "QuantitativeTrailLocus"        
#> [379] "r10kcod"                        "rae230a"                       
#> [381] "rae230b"                        "rat2302"                       
#> [383] "Rattus_norvegicus"              "Rattus_norvegicus_Data"        
#> [385] "Reactome"                       "Regression"                    
#> [387] "ReportWriting"                  "RepositoryData"                
#> [389] "ReproducibleResearch"           "ResearchField"                 
#> [391] "ResourceQueryingWorkflow"       "ReversePhaseProteinArray"      
#> [393] "ReversePhaseProteinArrayData"   "rgu34a"                        
#> [395] "rgu34b"                         "rgu34c"                        
#> [397] "rgug4130a"                      "ri16cod"                       
#> [399] "RiboSeq"                        "RIPSeq"                        
#> [401] "RIPSeqData"                     "RNASeq"                        
#> [403] "RNASeqData"                     "RNG_MRCChip"                   
#> [405] "rnu34"                          "Roberts2005Annotation"         
#> [407] "RocheChip"                      "rtu34"                         
#> [409] "rwgcod"                         "Saccharomyces_cerevisiae"      
#> [411] "Saccharomyces_cerevisiae_Data"  "Saccharum_officinarum"         
#> [413] "Saccharum_officinarum_Data"     "SAGE"                          
#> [415] "SAGEData"                       "SangerSeq"                     
#> [417] "SangerSeqData"                  "Scaffolding"                   
#> [419] "Schizosaccharomyces_pombe"      "Schizosaccharomyces_pombe_Data"
#> [421] "SequenceAnnotation"             "SequenceMatching"              
#> [423] "Sequencing"                     "SequencingData"                
#> [425] "SHDZ"                           "SIFT"                          
#> [427] "SingleCell"                     "SingleCellData"                
#> [429] "SingleCellWorkflow"             "SmallRNA"                      
#> [431] "SmallRNAData"                   "SNP"                           
#> [433] "SNPData"                        "SNPlocs"                       
#> [435] "Software"                       "Somatic"                       
#> [437] "SomaticMutation"                "SpecimenSource"                
#> [439] "SplicedAlignment"               "Staphylococcus_aureus"         
#> [441] "Staphylococcus_aureus_Data"     "StatisticalMethod"             
#> [443] "StemCell"                       "StructuralEquationModels"      
#> [445] "StructuralGenomics"             "StructuralPrediction"          
#> [447] "StructuralVariation"            "SupportVectorMachine"          
#> [449] "Survival"                       "Sus_scrofa"                    
#> [451] "Sus_scrofa_Data"                "SystemsBiology"                
#> [453] "Taeniopygia_guttata"            "Taeniopygia_guttata_Data"      
#> [455] "TargetedResequencing"           "Technology"                    
#> [457] "TechnologyData"                 "ThirdPartyClient"              
#> [459] "TimeCourse"                     "Tissue"                        
#> [461] "TissueMicroarray"               "TissueMicroarrayData"          
#> [463] "Toxoplasma_gondii"              "Transcription"                 
#> [465] "TranscriptomeVariant"           "Transcriptomics"               
#> [467] "Triticum_aestivum"              "Triticum_aestivum_Data"        
#> [469] "TwoChannel"                     "TwoChannelData"                
#> [471] "TxDb"                           "u133x3p"                       
#> [473] "UniversityHealthNetwork"        "VariantAnnotation"             
#> [475] "VariantDetection"               "Visualization"                 
#> [477] "Vitis_vinifera"                 "Vitis_vinifera_Data"           
#> [479] "WholeGenome"                    "Workflow"                      
#> [481] "WorkflowStep"                   "Xenopus_laevis"                
#> [483] "Xenopus_laevis_Data"            "Xenopus_tropicalis"            
#> [485] "Xenopus_tropicalis_Data"        "xenopuslaevis"                 
#> [487] "XtraSNPlocs"                    "yeast2"                        
#> [489] "ygs98"                          "Zea_mays"                      
#> [491] "Zea_mays_Data"                  "zebrafish"                     

params <- list(topics = c("Challenges", "Community"))
cat(format_topics(params))

#> - `Challenges`
#> 
#> - `Community`