Challenge Topics

Challenge Topics

topic_tags

format_topics(params)

Arguments

params
Challenge parameters as list.

Format

topic_tags is a character vector of valid topic tags.

Value

format_topics() returns a character value indicating the challenge topic tags.

Examples

topic_tags

#>   [1] "aCGH"                           "adme16cod"                     
#>   [3] "AffymetrixChip"                 "ag"                            
#>   [5] "AgilentChip"                    "Agroinformatics"               
#>   [7] "Alignment"                      "AlternativeSplicing"           
#>   [9] "Annotation"                     "AnnotationData"                
#>  [11] "AnnotationHub"                  "AnnotationHubSoftware"         
#>  [13] "AnnotationWorkflow"             "Anopheles_gambiae"             
#>  [15] "Anopheles_gambiae_Data"         "Apis_mellifera"                
#>  [17] "Apis_mellifera_Data"            "Arabidopsis_lyrata"            
#>  [19] "Arabidopsis_lyrata_Data"        "Arabidopsis_thaliana"          
#>  [21] "Arabidopsis_thaliana_Data"      "ArrayExpress"                  
#>  [23] "Asparagus_officinalis"          "AssayDomain"                   
#>  [25] "AssayDomainData"                "ATACSeq"                       
#>  [27] "ath1121501"                     "Bacillus_subtilis"             
#>  [29] "Bacillus_subtilis_Data"         "BasicWorkflow"                 
#>  [31] "BatchEffect"                    "Bayesian"                      
#>  [33] "BioCarta"                       "BiocViews"                     
#>  [35] "BiologicalQuestion"             "BiomedicalInformatics"         
#>  [37] "Biophysics"                     "BioPlex"                       
#>  [39] "Bos_taurus"                     "Bos_taurus_Data"               
#>  [41] "BreastCancerData"               "BSgenome"                      
#>  [43] "BugSigDB"                       "Caenorhabditis_elegans"        
#>  [45] "Caenorhabditis_elegans_Data"    "Callithrix_jacchus"            
#>  [47] "Callithrix_jacchus_Data"        "CancerData"                    
#>  [49] "Canis_familiaris"               "Canis_familiaris_Data"         
#>  [51] "cdf"                            "celegans"                      
#>  [53] "CellBasedAssays"                "CellBiology"                   
#>  [55] "CellCulture"                    "CGHData"                       
#>  [57] "Challenges"                     "Cheminformatics"               
#>  [59] "ChIPchip"                       "ChipDb"                        
#>  [61] "ChipManufacturer"               "ChipName"                      
#>  [63] "ChipOnChip"                     "ChipOnChipData"                
#>  [65] "ChIPSeq"                        "ChIPSeqData"                   
#>  [67] "Chlamydomonas_reinhardtii"      "Cicer_arietinum"               
#>  [69] "Ciona_intestinalis"             "Ciona_intestinalis_Data"       
#>  [71] "Classification"                 "ClonetechChip"                 
#>  [73] "Clustering"                     "CodelinkChip"                  
#>  [75] "ColonCancerData"                "Community"                     
#>  [77] "ComparativeGenomics"            "ComputationalChemistry"        
#>  [79] "COPDData"                       "CopyNumberVariation"           
#>  [81] "CopyNumberVariationData"        "Coverage"                      
#>  [83] "CpGIsland"                      "CpGIslandData"                 
#>  [85] "CRISPR"                         "CustomArray"                   
#>  [87] "CustomCDF"                      "CustomDBSchema"                
#>  [89] "Danio_rerio"                    "Danio_rerio_Data"              
#>  [91] "DataImport"                     "DataRepresentation"            
#>  [93] "db0"                            "ddPCR"                         
#>  [95] "DecisionTree"                   "DemethylateRegionDetection"    
#>  [97] "DenovoAssembler"                "DenovoGenome"                  
#>  [99] "DenovoTranscriptome"            "DifferentialDNA3DStructure"    
#> [101] "DifferentialExpression"         "DifferentialMethylation"       
#> [103] "DifferentialPeakCalling"        "DifferentialSplicing"          
#> [105] "DifferentialSplicingWorkflow"   "DimensionReduction"            
#> [107] "DiseaseModel"                   "DNA3DStructure"                
#> [109] "DNAMethylation"                 "DNASeq"                        
#> [111] "DNASeqData"                     "DNaseSeq"                      
#> [113] "DriverMutation"                 "drosgenome1"                   
#> [115] "Drosophila_melanogaster"        "Drosophila_melanogaster_Data"  
#> [117] "Drosophila_virilis"             "Drosophila_virilis_Data"       
#> [119] "drosophila2"                    "ENCODE"                        
#> [121] "Epigenetics"                    "EpigeneticsWorkflow"           
#> [123] "Epitranscriptomics"             "Eremothecium_gossypii"         
#> [125] "Eremothecium_gossypii_Data"     "Escherichia_coli"              
#> [127] "Escherichia_coli_Data"          "EuPathDB"                      
#> [129] "ExomeSeq"                       "ExomeSeqData"                  
#> [131] "ExonArray"                      "ExperimentalDesign"            
#> [133] "ExperimentData"                 "ExperimentHub"                 
#> [135] "ExperimentHubSoftware"          "ExpressionData"                
#> [137] "FeatureExtraction"              "FlowCytometry"                 
#> [139] "FlowCytometryData"              "FRMA"                          
#> [141] "FunctionalAnnotation"           "FunctionalGenomics"            
#> [143] "FunctionalPrediction"           "GACustomCDF"                   
#> [145] "Gallus_gallus"                  "Gallus_gallus_Data"            
#> [147] "Gasterosteus_aculeatus"         "Gasterosteus_aculeatus_Data"   
#> [149] "GEChip"                         "GeneCardsCustomSchema"         
#> [151] "GeneExpression"                 "GeneExpressionWorkflow"        
#> [153] "GeneFusionDetection"            "GenePrediction"                
#> [155] "GeneRegulation"                 "GeneSetEnrichment"             
#> [157] "GeneSignaling"                  "GeneTarget"                    
#> [159] "Genetics"                       "GeneticVariability"            
#> [161] "Genome"                         "GenomeAnnotation"              
#> [163] "GenomeAssembly"                 "GenomeBrowsers"                
#> [165] "GenomeWideAssociation"          "GenomicSequence"               
#> [167] "GenomicVariantsWorkflow"        "GenomicVariation"              
#> [169] "GenotypingArray"                "GenotypingArrayData"           
#> [171] "GEO"                            "Germline"                      
#> [173] "GermlineMutation"               "Glycine_max"                   
#> [175] "Glycine_max_Data"               "GO"                            
#> [177] "GraphAndNetwork"                "GUI"                           
#> [179] "h10kcod"                        "h20kcod"                       
#> [181] "HapMap"                         "hcg110"                        
#> [183] "hcgi12k"                        "hcgi8k"                        
#> [185] "hgfocus"                        "hgu133a"                       
#> [187] "hgu133a2"                       "hgu133b"                       
#> [189] "hgu133plus2"                    "hgu95a"                        
#> [191] "hgu95av2"                       "hgu95b"                        
#> [193] "hgu95c"                         "hgu95d"                        
#> [195] "hgu95e"                         "hguatlas13k"                   
#> [197] "hgug4100a"                      "hgug4101a"                     
#> [199] "hgug4110b"                      "hgug4111a"                     
#> [201] "hgug4112a"                      "hguqiagenv3"                   
#> [203] "hi16cod"                        "HiC"                           
#> [205] "HiddenMarkovModel"              "HighThroughputImagingData"     
#> [207] "HistoneModification"            "HIVData"                       
#> [209] "Homo_sapiens"                   "Homo_sapiens_Data"             
#> [211] "Hordeum_vulgare"                "Hordeum_vulgare_Data"          
#> [213] "hs25kresogen"                   "hu35ksuba"                     
#> [215] "hu35ksubb"                      "hu35ksubc"                     
#> [217] "hu35ksubd"                      "hu6800"                        
#> [219] "HuO22"                          "hwgcod"                        
#> [221] "IlluminaChip"                   "illuminaHumanv1"               
#> [223] "illuminaHumanv2"                "illuminaMousev1"               
#> [225] "illuminaMousev1p1"              "illuminaRatv1"                 
#> [227] "ImagingMassSpectrometry"        "ImagingMassSpectrometryData"   
#> [229] "ImmunoOncology"                 "ImmunoOncologyData"            
#> [231] "ImmunoOncologyWorkflow"         "indac"                         
#> [233] "INDACChip"                      "IndelDetection"                
#> [235] "Infrastructure"                 "InparanoidDb"                  
#> [237] "Instrumentation"                "JazaerimetaData"               
#> [239] "KEGG"                           "KidneyCancerData"              
#> [241] "Kluyveromyces_lactis"           "Kluyveromyces_lactis_Data"     
#> [243] "LeukemiaCancerData"             "LinkageDisequilibrium"         
#> [245] "Lipidomics"                     "lumiHumanV1"                   
#> [247] "lumiHumanV2"                    "lumiMouseV1"                   
#> [249] "lumiRatV1"                      "LungCancerData"                
#> [251] "m10kcod"                        "m20kcod"                       
#> [253] "Macaca_fascicularis"            "Macaca_mulatta"                
#> [255] "Macaca_mulatta_Data"            "Magnaporthe_grisea"            
#> [257] "Magnaporthe_grisea_Data"        "MassSpectrometry"              
#> [259] "MassSpectrometryData"           "MathematicalBiology"           
#> [261] "MBNICustomCDF"                  "Medicago_truncatul_Data"       
#> [263] "Medicago_truncatula"            "MeSHDb"                        
#> [265] "Metabolomics"                   "MetagenomeAssembly"            
#> [267] "Metagenomics"                   "MethylationArray"              
#> [269] "MethylationArrayData"           "MethylSeq"                     
#> [271] "MethylSeqData"                  "mgu74a"                        
#> [273] "mgu74av2"                       "mgu74b"                        
#> [275] "mgu74bv2"                       "mgu74c"                        
#> [277] "mgu74cv2"                       "mguatlas5k"                    
#> [279] "mgug4121a"                      "mgug4122a"                     
#> [281] "mi16cod"                        "Microarray"                    
#> [283] "MicroarrayData"                 "MicrobialStrain"               
#> [285] "Microbiome"                     "MicrobiomeData"                
#> [287] "MicroRNAArray"                  "MicroRNAArrayData"             
#> [289] "MicrosatelliteDetection"        "MicrotitrePlateAssay"          
#> [291] "MicrotitrePlateAssayData"       "miRNA"                         
#> [293] "miRNAData"                      "mm24kresogen"                  
#> [295] "MNaseSeq"                       "moe430a"                       
#> [297] "moe430b"                        "Monodelphis_domestica"         
#> [299] "Monodelphis_domestica_Data"     "MotifAnnotation"               
#> [301] "MotifDiscovery"                 "mouse4302"                     
#> [303] "mouse430a2"                     "mpedbarray"                    
#> [305] "mRNAArrayData"                  "mRNAMicroarray"                
#> [307] "mu11ksuba"                      "mu11ksubb"                     
#> [309] "Mu15v1"                         "mu19ksuba"                     
#> [311] "mu19ksubb"                      "mu19ksubc"                     
#> [313] "Mu22v3"                         "MultiChannel"                  
#> [315] "MultiChannelData"               "MultidimensionalScaling"       
#> [317] "MultipleComparison"             "MultipleSequenceAlignment"     
#> [319] "Mus_musculus"                   "Mus_musculus_Data"             
#> [321] "mwgcod"                         "NCI"                           
#> [323] "NCINatureCurated"               "Network"                       
#> [325] "NetworkEnrichment"              "NetworkInference"              
#> [327] "NeuralNetwork"                  "Neurospora_crassa"             
#> [329] "Neurospora_crassa_Data"         "Normalization"                 
#> [331] "Norway981"                      "NucleosomePositioning"         
#> [333] "Oncorhynchus_mykiss"            "Oncorhynchus_mykiss_Data"      
#> [335] "OneChannel"                     "OneChannelData"                
#> [337] "OperonHumanV3"                  "Organism"                      
#> [339] "OrganismData"                   "OrganismDb"                    
#> [341] "OrgDb"                          "Oryza_sativa"                  
#> [343] "Oryza_sativa_Data"              "OvarianCancerData"             
#> [345] "PackageType"                    "PackageTypeData"               
#> [347] "Pan_paniscus"                   "Pan_paniscus_Data"             
#> [349] "Pan_troglodytes"                "Pan_troglodytes_Data"          
#> [351] "PartheenMetaData"               "PathwayInteractionDatabase"    
#> [353] "Pathways"                       "PatternLogic"                  
#> [355] "PeakDetection"                  "pedbarrayv10"                  
#> [357] "pedbarrayv9"                    "Pharmacogenetics"              
#> [359] "Pharmacogenomics"               "Phylogenetics"                 
#> [361] "Plasmodium_falciparum"          "Plasmodium_falciparum_Data"    
#> [363] "PolyPhen"                       "PooledScreens"                 
#> [365] "Preprocessing"                  "PrincipalComponent"            
#> [367] "probe"                          "Project1000genomes"            
#> [369] "ProprietaryPlatforms"           "ProprietaryPlatformsData"      
#> [371] "ProstateCancerData"             "Proteome"                      
#> [373] "Proteomics"                     "ProteomicsWorkflow"            
#> [375] "Pseudomonas_aeruginosa"         "Pseudomonas_aeruginosa_Data"   
#> [377] "QiagenChip"                     "qPCR"                          
#> [379] "qPCRData"                       "QualityControl"                
#> [381] "QuantitativeTrailLocus"         "r10kcod"                       
#> [383] "rae230a"                        "rae230b"                       
#> [385] "rat2302"                        "Rattus_norvegicus"             
#> [387] "Rattus_norvegicus_Data"         "Reactome"                      
#> [389] "Regression"                     "ReportWriting"                 
#> [391] "RepositoryData"                 "ReproducibleResearch"          
#> [393] "ResearchField"                  "ResourceQueryingWorkflow"      
#> [395] "ReversePhaseProteinArray"       "ReversePhaseProteinArrayData"  
#> [397] "rgu34a"                         "rgu34b"                        
#> [399] "rgu34c"                         "rgug4130a"                     
#> [401] "ri16cod"                        "RiboSeq"                       
#> [403] "RIPSeq"                         "RIPSeqData"                    
#> [405] "RNASeq"                         "RNASeqData"                    
#> [407] "RNG_MRCChip"                    "rnu34"                         
#> [409] "Roberts2005Annotation"          "RocheChip"                     
#> [411] "rtu34"                          "rwgcod"                        
#> [413] "Saccharomyces_cerevisiae"       "Saccharomyces_cerevisiae_Data" 
#> [415] "Saccharum_officinarum"          "Saccharum_officinarum_Data"    
#> [417] "SAGE"                           "SAGEData"                      
#> [419] "SangerSeq"                      "SangerSeqData"                 
#> [421] "Scaffolding"                    "Schizosaccharomyces_pombe"     
#> [423] "Schizosaccharomyces_pombe_Data" "SequenceAnnotation"            
#> [425] "SequenceMatching"               "Sequencing"                    
#> [427] "SequencingData"                 "SHDZ"                          
#> [429] "SIFT"                           "SingleCell"                    
#> [431] "SingleCellData"                 "SingleCellWorkflow"            
#> [433] "SmallRNA"                       "SmallRNAData"                  
#> [435] "SNP"                            "SNPData"                       
#> [437] "SNPlocs"                        "Software"                      
#> [439] "Somatic"                        "SomaticMutation"               
#> [441] "Spatial"                        "SpatialData"                   
#> [443] "SpatialWorkflow"                "SpecimenSource"                
#> [445] "SplicedAlignment"               "Staphylococcus_aureus"         
#> [447] "Staphylococcus_aureus_Data"     "StatisticalMethod"             
#> [449] "StemCell"                       "StructuralEquationModels"      
#> [451] "StructuralGenomics"             "StructuralPrediction"          
#> [453] "StructuralVariation"            "SupportVectorMachine"          
#> [455] "Survival"                       "Sus_scrofa"                    
#> [457] "Sus_scrofa_Data"                "SystemsBiology"                
#> [459] "Taeniopygia_guttata"            "Taeniopygia_guttata_Data"      
#> [461] "TargetedResequencing"           "Technology"                    
#> [463] "TechnologyData"                 "ThirdPartyClient"              
#> [465] "TimeCourse"                     "Tissue"                        
#> [467] "TissueMicroarray"               "TissueMicroarrayData"          
#> [469] "Toxoplasma_gondii"              "Transcription"                 
#> [471] "TranscriptomeVariant"           "Transcriptomics"               
#> [473] "Triticum_aestivum"              "Triticum_aestivum_Data"        
#> [475] "TwoChannel"                     "TwoChannelData"                
#> [477] "TxDb"                           "u133x3p"                       
#> [479] "UniversityHealthNetwork"        "VariantAnnotation"             
#> [481] "VariantDetection"               "Visualization"                 
#> [483] "Vitis_vinifera"                 "Vitis_vinifera_Data"           
#> [485] "WholeGenome"                    "Workflow"                      
#> [487] "WorkflowStep"                   "Xenopus_laevis"                
#> [489] "Xenopus_laevis_Data"            "Xenopus_tropicalis"            
#> [491] "Xenopus_tropicalis_Data"        "xenopuslaevis"                 
#> [493] "XtraSNPlocs"                    "yeast2"                        
#> [495] "ygs98"                          "Zea_mays"                      
#> [497] "Zea_mays_Data"                  "zebrafish"                     

params <- list(topics = c("Challenges", "Community"))
cat(format_topics(params))

#> - `Challenges`
#> 
#> - `Community`